ocr/pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Keisuke Nakamura", email = "rauzi.univ@gmail.com" }
]
dependencies = [
    "mss>=10.0.0",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "opencv-python>=4.11.0.86",
    "numpy>=2.2.4",
    "flask>=3.1.0",
    "jsonrpcserver>=5.0.9",
    "pywin32>=310",
    "googletrans>=4.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ocr"]
